feat: Implement property highlighting feature and add announcement management routes
