+++ src/controllers/adminController.js
@@ -171,50 +171,3 @@ export const propertiesRegisteredMonthly = asyncHandler(async (req, res, next) =
     next(error);
   }
 });
-
-export const getAnnouncements = asyncHandler(async (req, res, next) => {
-  try {
-    const { active = null } = req.query;
-
-    const result = await AdminService.getAnnouncements(active);
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-export const getValidAnnouncements = asyncHandler(async (req, res, next) => {
-  try {
-    const result = await AdminService.getAnnouncements(true);
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-export const createAnnouncement = asyncHandler(async (req, res, next) => {
-  try {
-    const { data } = req.body;
-    const { photo } = req.body;
-
-    let announcementData = {};
-    if (data !== undefined) announcementData = JSON.parse(data);
-
-    const result = await AdminService.createAnnouncement(announcementData, photo);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-export const payAnnouncement = asyncHandler(async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const { transactionAmount, description, paymentMethod, payer } = req.body;
-
-    const result = await AdminService.payAnnouncement(id, transactionAmount, description, paymentMethod, payer);
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
