+++ src/controllers/propertyController.js
@@ -210,3 +210,15 @@ export const checkLimits = asyncHandler(async (req, res, next) => {
     next(error);
   }
 });
+
+export const highlightProperty = asyncHandler(async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const { email } = req;
+
+    const result = await PropertyService.highlightProperty(id, email);
+    res.status(200).json(result);
+  } catch (error) {
+    next(error);
+  }
+});
