+++ src/services/adminService.js
@@ -1,16 +1,12 @@
 import sgMail from '@sendgrid/mail';
 import dotenv from 'dotenv';
 import { initializeApp } from 'firebase/app';
-import { deleteObject, getDownloadURL, getStorage, ref, uploadBytesResumable } from 'firebase/storage';
-import { Payment } from 'mercadopago';
-import { v4 as uuid } from 'uuid';
+import { deleteObject, getStorage, ref } from 'firebase/storage';
 
-import { scheduleJob } from 'node-schedule';
 import firebaseConfig from '../config/firebase.js';
-import { checkIfPaid, paymentClient } from '../config/payment.js';
 import prisma from '../config/prisma.js';
 import ConfigurableError from '../errors/ConfigurableError.js';
-import { validateCpf, validateEmail, validatePrice, validateString } from '../validators/inputValidators.js';
+import { validateString } from '../validators/inputValidators.js';
 import PropertyService from './propertyService.js';
 import UserService from './userService.js';
 
@@ -248,139 +244,5 @@ export default class AdminService extends UserService {
     }
 
     return dataset;
-  }
-
-  static async getAnnouncements(active = null) {
-    const where = {};
-    if (active !== null) where.active = active;
-
-    return prisma.announcement.findMany({ where });
-  }
-
-  static async createAnnouncement(data, photo) {
-    if (!data) throw new ConfigurableError('Dados não informados', 400);
-    if (!['small', 'big'].includes(validateString(data.type))) throw new ConfigurableError('Tipo de anúncio inválido', 400);
-
-    if (!photo) throw new ConfigurableError('Foto não informada', 400);
-
-    // eslint-disable-next-line no-console
-    console.log(data);
-
-    const storageRef = ref(storage, `images/announcements/${validateEmail(data.announcerEmail)}-${photo.originalname}`);
-    const metadata = { contentType: photo.mimetype };
-    const snapshot = await uploadBytesResumable(storageRef, photo.buffer, metadata);
-    const downloadUrl = await getDownloadURL(snapshot.ref);
-
-    const announcementData = {
-      announcerName: validateString(data.announcerName),
-      announcerEmail: validateEmail(data.announcerEmail),
-      announcerCpf: validateCpf(data.announcerCpf),
-      photoUrl: downloadUrl,
-      siteUrl: validateString(data.siteUrl),
-      paymentType: 'pix', // at the moment, only pix is available
-      type: validateString(data.type),
-    };
-
-    const announcement = await prisma.announcement.create({ data: announcementData });
-
-    const modifier = announcement.type === 'small' ? 1 : 2;
-    const validPrices = [validatePrice('3.50') * modifier, validatePrice('7.50') * modifier, validatePrice('15.00') * modifier];
-
-    const payment = new Payment(paymentClient);
-
-    if (announcement.paymentType !== 'pix') throw new ConfigurableError('Método de pagamento inválido, no momento estamos aceitando apenas pix.', 400);
-    if (announcement.active) throw new ConfigurableError(`Este anúncio já está ativo, realize um novo pagamento quando ocorrer o vencimento em ${announcement.validUntil}.`, 400);
-
-    if (validPrices.includes(validatePrice(data.transactionAmount))) throw new ConfigurableError('O valor da transação deve ser maior que R$ 1,00.', 400);
-
-    const days = data.transactionAmount * (modifier === 1 ? 2 : 1);
-    const newValidDate = new Date(Date.now() + days * 24 * 60 * 60 * 1000);
-
-    const body = {
-      transaction_amount: validatePrice(data.transactionAmount),
-      description: validateString(data.description),
-      payment_method_id: 'pix',
-      payer: {
-        email: validateEmail(data.announcerEmail),
-        identification: {
-          type: 'CPF',
-          number: validateCpf(data.announcerCpf),
-        },
-      },
-    };
-
-    const requestOptions = { idempotencyKey: uuid() };
-
-    const result = await payment.create({ body, requestOptions });
-
-    scheduleJob({ start: new Date(), rule: '*/15 * * * * *' }, async () => {
-      if (checkIfPaid(result.id)) {
-        await prisma.announcement.update({ where: { id: announcement.id }, data: { active: true, validUntil: newValidDate } });
-      }
-    });
-
-    // eslint-disable-next-line no-console
-    // console.log(result);
-
-    const mailOptions = {
-      from: process.env.EMAIL_ADDRESS,
-      to: data.announcerEmail,
-      subject: 'Publicação de anúncio na Spaceimoveis',
-      html: `
-        <!DOCTYPE html>
-          <html>
-            <head>
-                <meta charset="UTF-8">
-                <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                <title>Pagamento - Space Imóveis</title>
-                <style>
-                    body {
-                        font-family: Arial, sans-serif;
-                        text-align: center;
-                        padding: 20px;
-                        background-color: #f4f4f4;
-                    }
-                    .container {
-                        max-width: 600px;
-                        background: white;
-                        padding: 20px;
-                        border-radius: 10px;
-                        box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-                        margin: 0 auto;
-                    }
-                    .btn {
-                        display: inline-block;
-                        padding: 10px 20px;
-                        font-size: 16px;
-                        color: white;
-                        background-color: #007BFF;
-                        text-decoration: none;
-                        border-radius: 5px;
-                        margin-top: 20px;
-                    }
-                    .btn:hover {
-                        background-color: #0056b3;
-                    }
-                </style>
-            </head>
-            <body>
-                <div class="container">
-                    <h2>Prezado(a) ${announcement.announcerName},</h2>
-                    <p>Esperamos que esta mensagem o encontre bem.</p>
-                    <p>Seu pagamento está pronto para ser processado. Para concluir a transação, clique no botão abaixo e acesse o link de pagamento da Space Imóveis:</p>
-                    <a href="${result.point_of_interaction.transaction_data.ticket_url}" class="btn">Realizar Pagamento</a>
-                    <p>Se tiver alguma dúvida, não hesite em nos contatar.</p>
-                    <p>Atenciosamente,</p>
-                    <p><strong>Space Imóveis</strong></p>
-                </div>
-            </body>
-          </html>
-      `,
-    };
-
-    let message = 'Anúncio criado com sucesso.';
-    sgMail.send(mailOptions).catch(() => { message += ' Mas o email não pode ser enviado.'; });
-
-    return { announcement, payment: result, message };
   }
 }
